MISRA 10_3 - majority of violations corrected
YT ID: DO-03

+++ lib/idtree.h
@@ -19,7 +19,7 @@
 #include "rb.h"
 
 
-#define MAX_ID ((1ULL << (__CHAR_BIT__ * (sizeof(int)) - 1)) - 1)
+#define MAX_ID ((1ULL << ((unsigned int)__CHAR_BIT__ * (sizeof(int)) - 1U)) - 1U)
 
 
 typedef rbtree_t idtree_t;
